# openapi.yaml
openapi: '3.0.2'
info:
  title: "Annotation tool"
  description: "The YAML API specification of the Annotation Tool, developed as part of the Software Project course."
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Local server
...
paths:
...
  /users:
    get:
      summary: Get all users
      description: Retrieve all users from the database
      tags:
        - Users
      responses:
        '200':
          description: "Successfully retrieved users"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Add a user
      description: Add a user to the database
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: "Successfully added the new user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: "Could not save user"
        '500':
          description: "Could not hash user's password"
  /users/{id}:
    get:
      summary: Get user
      description: Retrieve a user from the database
      tags:
        - Users
      parameters:
        - in: path
          name: "id"
          description: "ID of the user to be retrieved"
          required: true
          schema:
            type: string
            example: "supervisor@tudelft.nl"
      responses:
        '200':
          description: "Successfully retrieved user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '404':
          description: "Could not find the specified user"
    delete:
      summary: Delete user
      description: Delete a user from the database
      tags:
        - Users
        - User
      parameters:
        - in: path
          name: "id"
          description: "ID of the user to be deleted"
          required: true
          schema:
            type: string
            example: "supervisor@tudelft.nl"
      responses:
        '200':
          description: "Successfully deleted user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '404':
          description: "Could not find user"
...
  /users/updatePassword/{id}:
    put:
      summary: Update password
      description: Updates the specified user's password, in hashed form.
      tags:
        - Users
        - User
      parameters:
        - in: path
          name: "id"
          description: "ID of the user whose password should be updated"
          required: true
          schema:
            type: string
            example: "supervisor@tudelft.nl"
      responses:
        '200':
          description: "Successfully updated user's password"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: "Some field is empty"
        '403':
          description: "Old password does not match"
        '404':
          description: "User not found"
        '500':
          description: "Password hashing failed"
components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
          example: "supervisor@tudelft.nl"
        name:
          type: string
          example: "Supervisor One"
        password:
          type: string
          example: "superPass"
        role:
          type: string
          example: "supervisor"
        correspondingSubmissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionDB"
    SubmissionDB:
      type: object
      properties:
        email:
          type: string
          example: "student@tudelft.nl"
        file_submission:
          type: string
          format: binary
        group_name:
          type: string
          example: "test group"
        assignedCoordinators:
          type: array
          items:
            $ref: "#/components/schemas/User"
        file_name:
          type: string
          example: "submission.pdf"
        last_submitted:
          type: string
          example: "Tue, 10 Jan 2007 16:42:00 GMT"
        last_edited:
          type: string
          example: "Wed, 10 Mar 2006 08:12:22 GMT"
        is_submitted:
          type: boolean
          example: true
        is_locked:
          type: boolean
          example: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []